requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-jazzmin==2.6.0
sqlparse==0.4.3
tzdata==2022.7
